First half of the first test running.
I already consider splitting up this thing into multiple library files.
Also it might not be necessary to wrap the PayrollDatabase.employees
into an Option. I just did this because I thought it would be a nice way
to represent an empty database. In the end, at least at this stage, this
could also be done with an empty HashMap.